set `removeReflection` as core operation allow auto-removal of stale parents when running `removeReflection`
